requirements.txt
django
djangorestframework
Pillow
python-decouple
psycopg2-binary
python-slugify